+++ packages/react/src/atoms/labels/base.tsx
@@ -25,6 +25,8 @@ const LabelBase = (props: LabelBaseProps) => {
     theme
   );
 
+  // hiii
+
   return (
     <label
       htmlFor={props.htmlFor}