+++ packages/react/src/index.ts
@@ -101,6 +101,7 @@ export { default as DropdownBrowser } from "./molecules/dropdowns/browser";
 
 export { default as NavigationPageHeading } from "./molecules/navigations/heading/base";
 export { default as LazyLoadedImage } from "./molecules/lazyLoadedImage/Base";
+export { default as ImageOverlay } from "./molecules/ImageOverlay/base";
 
 //Slider
 
@@ -112,12 +113,16 @@ export { default as TabsBase } from "./molecules/tabs/base";
 //Centeralized Slider
 export { default as CentralizedSliderBase } from "./molecules/centralizedSlider/base";
 
+export { default as SingleDatetimePicker } from "./molecules/datetime/single";
+export { default as DateInput } from "./molecules/datetime/dateInput";
+export { default as DateInputExample } from "./molecules/datetime/dateInputExample";
+
 // ******************** End  ---- Molecules *************************
 
 // Start  ---- Organisms
 export { default as NavigationBase } from "./organisms/navigations/base";
 export { default as FooterBase } from "./organisms/footer/base";
-export { default as IconTitleDescriptionActionCard } from "./organisms/cards/IconTitleDescriptionActionCard/base"
+export { default as IconTitleDescriptionActionCard } from "./organisms/cards/IconTitleDescriptionActionCard/base";
 export { default as GalleryBase } from "./organisms/gallery/base";
 export { default as SectionContainerBase } from "./organisms/sectionContainer/base";
 export { default as ImageGallery } from "./organisms/imageGallery/base";